# video/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /douyin/favorite/count/user:
        get:
            tags:
                - FavoriteService
            description: 获取用户的获赞数和点赞数量
            operationId: FavoriteService_GetFavoriteCntByUId
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/favorite.api.favorite.v1.GetFavoriteCntByUIdResponse'
...
    /douyin/user:
        get:
            tags:
                - UserService
            description: |-
                获取用户信息，有 Redis，user + cnt
                 注意 is_follow 字段默认值，需要其他服务调用 favorite 服务取得
            operationId: UserService_UserInfo
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: string
                - name: token
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.UserResponse'
    /douyin/user/list:
        post:
            tags:
                - UserService
            description: 获取用户信息列表，使用用户 ID 列表查询
            operationId: UserService_UserInfoList
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.UserInfoListrRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.UserInfoListResponse'
    /douyin/user/login:
        post:
            tags:
                - UserService
            description: 用户登录，有 Redis
            operationId: UserService_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.UserLoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.UserLoginResponse'
    /douyin/user/register:
        post:
            tags:
                - UserService
            description: 用户注册
            operationId: UserService_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.UserRegisterRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.UserRegisterResponse'
    /douyin/user/update_favorite_count:
        post:
            tags:
                - UserService
            description: |-
                更新点赞数量
                 FavoriteCount 和 TotalFavorited
            operationId: UserService_UpdateFavoriteCnt
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.UpdateFavoriteCntRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.UpdateFavoriteCntResponse'
    /douyin/user/update_follow_count:
        post:
            tags:
                - UserService
            description: |-
                更新关注数量
                 FollowCount 和 FollowerCount
            operationId: UserService_UpdateFollowCnt
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.UpdateFollowCntRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.UpdateFollowCntResponse'
    /douyin/user/update_work_count:
        post:
            tags:
                - UserService
            description: |-
                更新计数
                 更新作品数量
            operationId: UserService_UpdateWorkCnt
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.UpdateWorkCntRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.UpdateWorkCntResponse'
...
components:
    schemas:
...
        favorite.api.favorite.v1.GetFavoriteCntByUIdResponse:
            type: object
            properties:
                totalFavorited:
                    type: string
                favoriteCount:
                    type: string
...
        user.v1.UpdateFavoriteCntRequest:
            type: object
            properties:
                favoriteUserId:
                    type: string
                favoritedUserId:
                    type: string
                favoriteCount:
                    type: string
                totalFavorited:
                    type: string
        user.v1.UpdateFavoriteCntResponse:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
        user.v1.UpdateFollowCntRequest:
            type: object
            properties:
                followUserId:
                    type: string
                followedUserId:
                    type: string
                followCount:
                    type: string
                followerCount:
                    type: string
        user.v1.UpdateFollowCntResponse:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
        user.v1.UpdateWorkCntRequest:
            type: object
            properties:
                userId:
                    type: string
                workCount:
                    type: string
        user.v1.UpdateWorkCntResponse:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
        user.v1.User:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                followCount:
                    type: string
                followerCount:
                    type: string
                isFollow:
                    type: boolean
                avatar:
                    type: string
                backgroundImage:
                    type: string
                signature:
                    type: string
                totalFavorited:
                    type: string
                workCount:
                    type: string
                favoriteCount:
                    type: string
        user.v1.UserInfoListResponse:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
                users:
                    type: array
                    items:
                        $ref: '#/components/schemas/user.v1.User'
        user.v1.UserInfoListrRequest:
            type: object
            properties:
                userId:
                    type: array
                    items:
                        type: string
        user.v1.UserLoginRequest:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
        user.v1.UserLoginResponse:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
                userId:
                    type: string
                token:
                    type: string
        user.v1.UserRegisterRequest:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
        user.v1.UserRegisterResponse:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
                userId:
                    type: string
                token:
                    type: string
        user.v1.UserResponse:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
                user:
                    $ref: '#/components/schemas/user.v1.User'
...